Work_JO/CP1/requirements.txt
cryptocraphy == 39.0.2
b64uuid == 0.1
boto3 == 1.26.94
python-dotenv == 1.0.0
APScheduler == 3.10.1